+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.359](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.358...v3.0.0-beta.359) (2024-02-26)
+
+
+### Bug Fixes
+
+* reject guest token downscope ([#3398](https://github.com/webex/webex-js-sdk/issues/3398)) ([ae70cf0](https://github.com/webex/webex-js-sdk/commit/ae70cf0bdfc78b4a2b31f541c2e6da1ff5a3ffa5))
+
+## [3.0.0-beta.358](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.357...v3.0.0-beta.358) (2024-02-23)
+
+
+### Bug Fixes
+
+* **plugin-meetings:** update MQE data related to hop-by-hop loss ([#3388](https://github.com/webex/webex-js-sdk/issues/3388)) ([9b8457c](https://github.com/webex/webex-js-sdk/commit/9b8457c947fa4a2021cb9b15cba859c2fa0fc67e))
+
+## [3.0.0-beta.357](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.356...v3.0.0-beta.357) (2024-02-23)
+
+
+### Bug Fixes
+
+* **test:** remove only tag for test ([#3399](https://github.com/webex/webex-js-sdk/issues/3399)) ([d0c8390](https://github.com/webex/webex-js-sdk/commit/d0c8390b4d0440d11b51c46b6f08eb142fae7022))
+
 ## [3.0.0-beta.356](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.355...v3.0.0-beta.356) (2024-02-23)